+++ lib/cpp/src/concurrency/ThreadManager.h
@@ -127,7 +127,7 @@ class ThreadManager {
    *
    * @throws TooManyPendingTasksException Pending task count exceeds max pending task count
    */
-  virtual void add(boost::shared_ptr<Runnable>task, long long timeout=0LL) = 0;
+  virtual void add(boost::shared_ptr<Runnable>task, int64_t timeout=0LL) = 0;
 
   /**
    * Removes a pending task
